implement check for system variable on windows before creating it
